improve memory usage and switch to libvips for slight perf boost
Using xfades in an ffmpeg pipeline doesn't close the previous
filters until the whole video is gone. This effectively causes
this filter to leak memory due to the cached masks, which is an
issue if it is used with many videos. To curb this, two measures
were taken:
- use limited LRU cache instead of unbounded one
- drop the cache if no images were processed for some time
This won't be perfect, but hopefully good enough to stop renders
from using 20G+ of memory.

Additionally this PR switches to libvips for a tiny real time
speed gain, though the CPU used stays unchanged. The blurs are
also slightly different due to this.